fix(kmp/phase-3): cheapshark storeID is a JSON string, not int
The cheapshark `/api/1.0/stores` endpoint returns `"storeID": "1"`
(string) but RemoteStore typed it as `Int`. kotlinx-serialization is
strict about numeric coercion: a string-into-Int field throws
SerializationException.

Symptoms in production:
- Ktor REQUEST line logged
- No Ktor RESPONSE line, because the response body parsing throws
- StoresApi's try/catch catches the SerializationException as
  ApiResponse.exception(t), but the user observes no data on the
  stores screen and no error feedback (the error UI for stores was
  presumably never tested with a real failure mode).

The previous Retrofit + kotlinx-retrofit-converter setup must have
been more lenient — possibly the cheapshark API recently changed,
or the converter had quiet coercion. Either way, Ktor's
ContentNegotiation surfaces the strict behaviour.

Fix:
- RemoteStore.storeID -> String.
- StoreMappers.toStore() does .toInt() before populating Store.
- Test fixture STORE_LIST_BODY updated to "storeID": "1" so it
  matches the real API shape.

Removes the temporary diagnostic Log.d bracketing in fetchStores
that was added in 8d0e67c — `android.util.Log` isn't mocked in JVM
tests so it broke the test suite, and the bug is now identified
without needing the diagnostic.

Keeps the `requestTimeoutMillis = 30_000` from b2c4aa3 — useful
defensive setting against future read-side hangs even though it
wasn't load-bearing for this particular bug.

+++ remote/cheapshark/src/androidMain/kotlin/pm/bam/gamedeals/remote/cheapshark/mappers/StoreMappers.kt
@@ -14,7 +14,7 @@ internal fun RemoteStore.RemoteStoreImages.toStoreImages(): Store.StoreImages =
 
 internal fun RemoteStore.toStore(): Store =
     Store(
-        storeID = storeID,
+        storeID = storeID.toInt(),
         storeName = storeName,
         isActive = isActive.toBooleanStrict(),
         images = images.toStoreImages()
